LPS-94745 Move localization logic to SPI
