+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/field/type/v1_0/CheckboxFieldType.java
@@ -63,7 +63,7 @@ public class CheckboxFieldType extends BaseFieldType {
 			"showAsSwitcher", jsonObject.getBoolean("showAsSwitcher"));
 
 		spiDataDefinitionField.setDefaultValue(
-			LocalizedValueUtil.toLocalizationMap(
+			LocalizedValueUtil.toLocalizedValues(
 				jsonObject.getJSONObject("predefinedValue")));
 
 		return spiDataDefinitionField;
