+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/field/type/v1_0/CheckboxMultipleFieldType.java
@@ -68,7 +68,7 @@ public class CheckboxMultipleFieldType extends BaseFieldType {
 			"showAsSwitcher", jsonObject.getBoolean("showAsSwitcher"));
 
 		spiDataDefinitionField.setDefaultValue(
-			LocalizedValueUtil.toLocalizationMap(
+			LocalizedValueUtil.toLocalizedValues(
 				jsonObject.getJSONObject("predefinedValue")));
 
 		return spiDataDefinitionField;
