+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/field/type/v1_0/DateFieldType.java
@@ -56,7 +56,7 @@ public class DateFieldType extends BaseFieldType {
 			jsonObject);
 
 		spiDataDefinitionField.setDefaultValue(
-			LocalizedValueUtil.toLocalizationMap(
+			LocalizedValueUtil.toLocalizedValues(
 				jsonObject.getJSONObject("predefinedValue")));
 
 		return spiDataDefinitionField;
