+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/field/type/v1_0/EditorFieldType.java
@@ -56,7 +56,7 @@ public class EditorFieldType extends BaseFieldType {
 
 		customProperties.put(
 			"placeholder",
-			LocalizedValueUtil.toLocalizationMap(
+			LocalizedValueUtil.toLocalizedValues(
 				jsonObject.getJSONObject("placeholder")));
 
 		return spiDataDefinitionField;
