+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/field/type/v1_0/ParagraphFieldType.java
@@ -59,7 +59,7 @@ public class ParagraphFieldType extends BaseFieldType {
 
 		customProperties.put(
 			"text",
-			LocalizedValueUtil.toLocalizationMap(
+			LocalizedValueUtil.toLocalizedValues(
 				jsonObject.getJSONObject("text")));
 
 		return spiDataDefinitionField;
