+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/field/type/v1_0/RadioFieldType.java
@@ -70,7 +70,7 @@ public class RadioFieldType extends BaseFieldType {
 				jsonObject.getJSONObject("options")));
 		customProperties.put(
 			"predefinedValue",
-			LocalizedValueUtil.toLocalizationMap(
+			LocalizedValueUtil.toLocalizedValues(
 				jsonObject.getJSONObject("predefinedValue")));
 
 		return spiDataDefinitionField;
