+++ modules/apps/data-engine/data-engine-rest-impl/src/main/java/com/liferay/data/engine/rest/internal/field/type/v1_0/SelectFieldType.java
@@ -69,7 +69,7 @@ public class SelectFieldType extends BaseFieldType {
 				jsonObject.getJSONObject("options")));
 		customProperties.put(
 			"predefinedValue",
-			LocalizedValueUtil.toLocalizationMap(
+			LocalizedValueUtil.toLocalizedValues(
 				jsonObject.getJSONObject("predefinedValue")));
 
 		return spiDataDefinitionField;
